Add SuccessfulOutputsExecutionRecorder
